API Freelance/requirements.txt
fastapi
uvicorn[standard]
PyJWT
python-multipart
email-validator